Remove circular dependencies, remove iron references, remove polymer references, update .circleci, update docker files, update express.js, remove redundant fiels from root, update rollup for build, update.index.html
