+++ .circleci/config.yml
@@ -4,7 +4,7 @@ jobs:
     working_directory: ~/code
     # The primary container is an instance of the first list image listed. Your build commands run in this container.
     docker:
-      - image: node:12.22.7
+      - image: node:14.21
     steps:
       - checkout
       - restore_cache:
@@ -17,14 +17,14 @@ jobs:
           paths:
             - node_modules
       - run:
-          name: Polymer3 Build
-          command: npm run build
+          name: EsLint
+          command: npm run lint
 
   build_and_deploy:
     working_directory: ~/code
     # The primary container is an instance of the first list image listed. Your build commands run in this container.
     docker:
-      - image: node:12.22.7
+      - image: node:14.21
     steps:
       - checkout
       - setup_remote_docker
@@ -38,6 +38,7 @@ jobs:
             sed -i "0,/revNo/s//$REVNO/" version.json
             sed -i "0,/bDate/s//$BUILDDATE/" package.json
             sed -i "0,/bDate/s//$BUILDDATE/" version.json
+            sed -i "0,/bDate/s//$BUILDDATE/" index.html
       - run:
           name: Install Docker client
           command: |
@@ -73,3 +74,4 @@ workflows:
                 - master
                 - staging
                 - develop
+                - shoelace-migration
