+++ Dockerfile-dev
@@ -3,7 +3,6 @@ RUN apk update
 RUN apk add --update bash
 
 RUN apk add git
-RUN npm install -g --unsafe-perm polymer-cli
 RUN npm install -g gulp
 RUN npm install -g typescript
