+++ package.json
@@ -26,21 +26,16 @@
   },
   "dependencies": {
     "@mapbox/leaflet-omnivore": "^0.3.4",
-    "@polymer/app-localize-behavior": "^3.0.1",
-    "@polymer/iron-autogrow-textarea": "^3.0.3",
-    "@polymer/paper-listbox": "^3.0.1",
-    "@polymer/paper-tabs": "^3.1.0",
-    "@polymer/polymer": "^3.5.1",
     "@shoelace-style/shoelace": "^2.10.0",
     "@types/leaflet": "^1.9.3",
     "@types/lodash-es": "^4.17.7",
     "@types/ramda": "0.27.37",
-    "@unicef-polymer/etools-form-builder": "^3.0.0-rc.4",
+    "@unicef-polymer/etools-form-builder": "^3.0.0-rc.5",
     "@unicef-polymer/etools-info-tooltip": "^4.1.11",
-    "@unicef-polymer/etools-modules-common": "^5.0.0-rc.26",
+    "@unicef-polymer/etools-modules-common": "^5.0.0-rc.27",
     "@unicef-polymer/etools-piwik-analytics": "^4.0.1-rc.3",
     "@unicef-polymer/etools-types": "^1.11.16",
-    "@unicef-polymer/etools-unicef": "^1.0.0-rc.34",
+    "@unicef-polymer/etools-unicef": "^1.0.0-rc.35",
     "@unicef-polymer/etools-utils": "^2.0.0-rc.3",
     "@webcomponents/webcomponentsjs": "^2.8.0",
     "dayjs": "^1.11.7",
@@ -87,7 +82,6 @@
     "mout": ">=1.2.4",
     "node-fetch": ">=2.6.7",
     "path-parse": ">=1.0.7",
-    "polymer-cli": "^1.9.11",
     "prettier": "~2.5.1",
     "qs": ">=6.5.3",
     "rollup": "^2.79.1",
