+++ src_ts/components/pages/activities-and-data-collection/activity-item/details-tab/details-cards/entities-monitor-card/entities-monitor-card.ts
@@ -1,6 +1,5 @@
-import {css, TemplateResult, CSSResultArray} from 'lit';
+import {css, TemplateResult, html, CSSResultArray} from 'lit';
 import {customElement, property} from 'lit/decorators.js';
-import {template} from './entities-monitor-card.tpl';
 import {elevationStyles} from '../../../../../../styles/elevation-styles';
 import {SharedStyles} from '../../../../../../styles/shared-styles';
 import {BaseDetailsCard} from '../base-details-card';
@@ -17,8 +16,10 @@ import {FlexLayoutClasses} from '../../../../../../styles/flex-layout-classes';
 import {openDialog} from '@unicef-polymer/etools-utils/dist/dialog.util';
 import {simplifyValue} from '../../../../../../utils/objects-diff';
 import {InterventionsMixin} from '../../../../../../common/mixins/interventions-mixin';
+import {translate} from 'lit-translate';
 
 export const CARD_NAME = 'entities-monitor';
+const ELEMENT_FIELDS: (keyof IActivityDetails)[] = ['cp_outputs', 'partners', 'interventions'];
 
 @customElement('entities-monitor-card')
 export class EntitiesMonitorCard extends InterventionsMixin(PartnersMixin(CpOutputsMixin(BaseDetailsCard))) {
@@ -34,7 +35,71 @@ export class EntitiesMonitorCard extends InterventionsMixin(PartnersMixin(CpOutp
   }
 
   render(): TemplateResult {
-    return template.call(this);
+    return html`
+      <etools-card
+        card-title="${translate('ACTIVITY_DETAILS.ENTRIES_TO_MONITOR')}"
+        ?is-editable="${this.havePossibilityToEditCard(CARD_NAME, ELEMENT_FIELDS)}"
+        ?edit="${this.isEditMode}"
+        @start-edit="${() => this.startEdit()}"
+        @save="${() => this.save()}"
+        @cancel="${() => this.cancel()}"
+      >
+        <div class="card-content" slot="content">
+          <!--   Spinner for loading data   -->
+          <etools-loading
+            ?active="${this.isLoad}"
+            loading-text="${translate('MAIN.LOADING_DATA_IN_PROCESS')}"
+          ></etools-loading>
+
+          <!--   Spinner for updating data   -->
+          <etools-loading
+            ?active="${this.isUpdate}"
+            loading-text="${translate('MAIN.SAVING_DATA_IN_PROCESS')}"
+          ></etools-loading>
+
+          <div class="layout horizontal">
+            <!--    Partners List    -->
+            <entries-list
+              class="entries-list"
+              .nameList="${translate('ACTIVITY_DETAILS.PARTNERS')}"
+              .formatItem="${(item: EtoolsPartner) => item.name}"
+              .items="${this.activityPartners}"
+              ?is-readonly="${!this.isEditMode}"
+              @add-entry="${() => this.openAddPartner()}"
+              @remove-entry="${({detail}: CustomEvent) =>
+                this.removeItem<EtoolsPartner>(detail.id, 'partners', 'activityPartners')}"
+            >
+            </entries-list>
+
+            <!--    CP Outputs List    -->
+            <entries-list
+              class="entries-list"
+              .nameList="${translate('ACTIVITY_DETAILS.CP_OUTPUTS')}"
+              .formatItem="${(item: EtoolsCpOutput) => item.name}"
+              .items="${this.activityCpOutputs}"
+              ?is-readonly="${!this.isEditMode}"
+              @add-entry="${() => this.openAddCpOutput()}"
+              @remove-entry="${({detail}: CustomEvent) =>
+                this.removeItem<EtoolsCpOutput>(detail.id, 'cp_outputs', 'activityCpOutputs')}"
+            >
+            </entries-list>
+
+            <!--    Interventions List    -->
+            <entries-list
+              class="entries-list"
+              .nameList="${translate('ACTIVITY_DETAILS.INTERVENTIONS')}"
+              .formatItem="${(item: EtoolsIntervention) => item.title}"
+              .items="${this.activityInterventions}"
+              ?is-readonly="${!this.isEditMode}"
+              @add-entry="${() => this.openAddIntervention()}"
+              @remove-entry="${({detail}: CustomEvent) =>
+                this.removeItem<EtoolsIntervention>(detail.id, 'interventions', 'activityInterventions')}"
+            >
+            </entries-list>
+          </div>
+        </div>
+      </etools-card>
+    `;
   }
 
   protected startEdit(): void {
