+++ src_ts/components/pages/analyze/monitoring-tab/geographic-coverage/geographic-coverage.tpl.ts
@@ -52,8 +52,8 @@ export function template(this: GeographicCoverageComponent): TemplateResult {
                 trigger-value-change-event
                 @etools-selected-items-changed="${({detail}: CustomEvent) =>
                   this.onSelectionChange(detail.selectedItems)}"
-                @iron-overlay-closed="${() => this.onDropdownClose()}"
-                @remove-selected-item="${(event: Event) => this.onRemoveSelectedItem(event)}"
+                @dropdown-closed="${() => this.onDropdownClose()}"
+                @removed-selected-items="${(event: Event) => this.onRemoveSelectedItems(event)}"
                 hide-search
                 allow-outside-scroll
               >
