+++ src_ts/components/pages/templates/issue-tracker-tab/issue-tracker-tab-statuses.ts
@@ -1,0 +1,6 @@
+import {get as getTranslation} from 'lit-translate';
+
+export const ISSUE_STATUSES: DefaultDropdownOption<string>[] = [
+  {value: 'new', display_name: getTranslation('ISSUE_TRACKER.STATUSES.NEW')},
+  {value: 'past', display_name: getTranslation('ISSUE_TRACKER.STATUSES.PAST')}
+];
